==> 00-load-libraries.R <==
# 00-load-libraries.R
# Loads all libraries used in SNAPP-CRF
# case study analysis
# W. Friedman 2/7/2022

library(here)
library(tidyverse)
library(ggplot2)
library(RColorBrewer)
library(janitor)
library(googlesheets4)
library(kableExtra)
library(broom)
library(ggpubr)
library(psych)
library(corrplot)
library(cluster)
library(viridis)
library(tidytext)
