==> 01-prep-attribute-tables.Rmd <==
---
title: "01-read-attribs.Rmd"
author: "Whitney Friedman"
date: "2/4/2022"
output: html_document
---

```{r setup, include=FALSE, message = FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

#Load libraries
source(here("00-load-libraries.R"))
```


# Read in the data from Google Sheets
```{r, results=FALSE, message = FALSE, warning = FALSE, class.source = 'fold-hide', warning = FALSE}
aguion_barnacle <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1WfGsKOT0gFSN1yHUObDNSQflqd5avodSuf3kOZLTyLk/edit#gid=1947230832", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Aguion",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))


#######################################################################################################################
burden_sardine <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1TkB-P94xgRssmcbAuzwoh0BlDdvUfYHO7vWjK6r2pxs/edit#gid=1965519914", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names()%>% 
  add_column(case_study = "Burden",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))


#######################################################################################################################
dickey_collas_eur <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1hfXTM5hwe0iLWQQmrc31-xruOV3SUixuxqRSuqC8ydA/edit#gid=1707192308", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Dickey_Collas",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



#######################################################################################################################
eurich_clam <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1ZEIev1lcoD6Lxu3otzDCfbJJSiPJoZeyYbTvBXb_S9U/edit#gid=1453404713", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Eurich",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



#####################################################################################################################################
free_crab <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/12zMb5el0QAB-CdZdmdKjWw5Krbwx7farqlUvbN08HEg/edit#gid=390317686", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Free",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



#######################################################################################################################
golden_madagascar <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1w99nN8C_Vd82j4Y23g2ZG_kuEh_Jc2HkW6lnLkhzeCU/edit#gid=390607640", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Golden",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference)) %>% 
  mutate(resilience_attribute = recode(resilience_attribute, 
                                       "Flexible and agile infrastructure" = 
                                       "Flexible and agile infrstructure")) # UPDATE MEGHAN (remove)

golden_madagascar$resilience_attribute

#######################################################################################################################
friedman_fiji <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/13fgkP1xHZB6gZqXUBHuylaC3ImCDIDTFFjqQ8RAoV-s/edit#gid=476628697", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Friedman",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



#######################################################################################################################
hollowed_bering_sea <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1YOYZ9T7RR_8OkqQIWeZobhS-3yaEbKmuAVz16Nrrz9g/edit#gid=9514874", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Hollowed",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



########################################################################################################################
kisara_squid <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1RX1iUmEKsOUmuPZkn4SMA6zts9TTPz-gi0D6XFrY1RQ/edit#gid=934077877", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Kisara",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



###########################################################################################################################
kleisner_jfi <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1YM1wc6F0SgGIoBcTvSjVyLJquE6-6Chk1dPFoLiXuQY/edit#gid=1865610127", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Kleisner",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



################################################################################################################################
lau_png <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/18po7Pgq5Od6Q8yCnFdSfj_Gaxii34e-WsYwBFA3RTvo/edit#gid=1980661832", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Lau",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



###############################################################################################################################
mason_iceland <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1ZPs5ow4wpZgj5eP26jvoMRgD9q6VT9DEdkJfqj7UrRE/edit#gid=2142384281", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Mason",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))




##########################################################################################################################
mills_lobster <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1HEPT54TsR1aTNNzDtZI59zU0eb-eJfH61ujErxAFtiI/edit#gid=553950990", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Mills",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



########################################################################################################################
pecl_lobster <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/11Z-SLlAS0f9Ug1Z8gLMnHkaN7hoZT1W0KMVLG8HtbXs/edit#gid=179679195", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Pecl",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



###########################################################################################################################
#schmidt_senegal <- read_sheet(ss = #"https://docs.google.com/spreadsheets/d/1bUY6yII5y0FZfDRJ5lxpA7iD2p3o__2A1xsWivxwKB0/edit#gid=2000123802", 
#                           sheet = "UPDATED_S4_ResilienceAttributes",
#                           range = "A6:P61",
#                           col_names = TRUE,
#                           na = c("NA","na","NAN","nan",""),
#                           guess_max = 10000,
#                           trim_ws = TRUE,
#                           col_types = "c"
#                           ) %>%
#  as_tibble() %>%
#  clean_names() %>% 
#  add_column(case_study = "Schmidt",
#             reference = seq(1:nrow(.)),
#             .after = "question_id") %>% 
#  mutate(reference = str_c(case_study, reference))



###############################################################################################################################
tokunaga_lobster <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1jgFZut693B14rdKbxprKtQcoh-mFlMzAktekjhvwf2Q/edit#gid=228934601", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Tokunaga",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



############################################################################################################################
westfall_hms <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1DMx0Ux9Li1LhIe4ee5bbAYtJaw1vCezziRGyeERBthY/edit#gid=1818956842", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Westfall",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))



###################################################################################################################################
zhao_moorea <- read_sheet(ss = "https://docs.google.com/spreadsheets/d/1-Muz76--fNv1qfdXZcBENP-agemTTySZAay9uAEO0HI/edit#gid=1980661832", 
                           sheet = "UPDATED_S4_ResilienceAttributes",
                           range = "A6:P61",
                           col_names = TRUE,
                           na = c("NA","na","NAN","nan",""),
                           guess_max = 10000,
                           trim_ws = TRUE,
                           col_types = "c"
                           ) %>%
  as_tibble() %>%
  clean_names() %>% 
  add_column(case_study = "Zhao",
             reference = seq(1:nrow(.)),
             .after = "question_id") %>% 
  mutate(reference = str_c(case_study, reference))
```


##############################################################################################

# Merge the datasets

```{r}
attributes_merged <- rbind(aguion_barnacle, burden_sardine, dickey_collas_eur, eurich_clam, free_crab, friedman_fiji, golden_madagascar, hollowed_bering_sea, kisara_squid, kleisner_jfi, lau_png, mason_iceland, mills_lobster, pecl_lobster, tokunaga_lobster, westfall_hms, zhao_moorea)

```


# Clean the merged datasets 

```{r}
attributes_merged_clean <- attributes_merged %>% 
  rename(dimension=dimensions, 
         domain=new_domain, 
         attribute=resilience_attribute, 
         score=score_1_4,
         quality=information_quality, 
         importance=importance_of_attribute_in_study_system_high_medium_low, 
         work_as_described=does_the_mechanism_work_as_described_in_this_fishery_system) %>% 
# Simplify
  select(dimension, domain, case_study, reference, attribute, score, quality, importance, work_as_described) %>% 
# Remove spacer rows
  filter(!is.na(attribute)) %>% 
# Format scores
  filter(!is.na(score)) %>% 
# In order to get rid of "NA" values that were purposefully entered into the data, I used the `case_when` function to revert
  mutate(official_score = case_when(
    score == "1" ~ "1",
    score == "2" ~ "2",
    score == "3" ~ "3",
    score == "4" ~ "4",
    score == "NA" ~ "0"
  )) %>% 
  mutate(official_score=as.numeric(official_score)) %>%
  # Formate dimension
  mutate(dimension=recode_factor(dimension, 
                                 "Ecological"="Ecological",
                                 "Socio-economic"="Social-economic",
                                 "Governance-management"="Goverance")) %>% 
# Format data quality
  mutate(quality=ifelse(quality=="NA - Not relevant in this system", NA, quality),
         quality=recode_factor(quality,
                               "E - No data/information; no basis for expert judgement"="No data", ### CF used "E - No data/information; no basis for expert judgement", but JE's was different
                               "D - not confident that the answer provided reflects the true state of the system"="Low",
                               "C - fairly confident that the answer provided reflects the true state of the system"="Fair",
                               "B - limited data/information and expert judgement"="Good",
                               "A - adequate and reliable data/information"="Excellent")) %>% 
    # To fix some errors around data quality in the original datasets
  mutate(quality=recode_factor(quality,
                               "E - No data"="No data")) %>% 
  # Create a new column to house numeric values for information quality
  mutate(numeric_quality = case_when(
    quality == "Excellent" ~ 4,
    quality == "Good" ~ 3,
    quality == "Fair" ~ 2,
    quality == "Low" ~ 1,
    quality == "No data" ~ 0)) %>% 
  # Burden has NA's across this seciton an can't be in data (Take out the # when you are working on scores and importance aadd bakc in when working in data quality until this is updated in his sheet)
  #filter(!is.na(numeric_quality)) %>%
  # Format importance
  mutate(importance=stringr::str_to_sentence(importance),
         importance=factor(importance, levels=c("Low", "Medium", "High"))) %>% 
  # Create a new column to house numeric values for importance
  mutate(numeric_importance = case_when(
    importance == "Low" ~ "1",
    importance == "Medium" ~ "2",
    importance == "High" ~ "3")) %>% 
  mutate(numeric_importance=as.numeric(numeric_importance)) %>% 
  filter(!is.na(numeric_importance))
  

```

# Create final table for export
```{r}
attribs <- attributes_merged_clean %>% 
  mutate(attribute_id = str_c(dimension, attribute, sep = "_"),
         attribute_id = str_replace_all(attribute_id,"[:blank:]|[:punct:]","_")) %>% 
  select(-score) %>% 
  rename(score = official_score)

attribs
```


# Save
Save workspace image & csv to load with other scripts
```{r}

save(attribs, file = here("data","attribs.Rdata"))
attribs %>% write_csv(here("data","attribs.csv"))

save(attributes_merged_clean, file = here("data","attributes_merged_clean.Rdata"))
attributes_merged_clean %>% write_csv(here("data","attributes_merged_clean.csv"))

```
