==> 04-attribute-score-stats.Rmd <==
---
title: "Eavlauting Attribute Scores Across Cases"
author: "Meghan Fletcher"
date: "1/31/2022"
output: 
  html_document:
    theme: flatly
    code_folding: hide
    toc: true
    toc_float: true
---

```{r setup, include=FALSE, message = FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

```

# Load data

```{r}
library(here)
source(here("00-load-libraries.R"))
load(here("data","attribute_table_clean.Rdata")) # from 01-prep-attribute-tables.Rmd
load(here("data","attributes_merged_clean.Rdata"))

```

############################################################################################

# Find Attribute Score Means 
The attribute score means of each attribute are reflected in the table. These means are based off of all case studies combined.

```{r}
# Build a table
merged_mean <- attributes_merged_clean %>% 
  group_by(attribute) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Score = score,
         Attribute = attribute) %>% 
  mutate(across(where(is.numeric), ~ round(., 3)))

# Use `kableExtra` package to create table
merged_mean %>% 
  kbl(caption = "Resilience Attribute Score Means") %>% 
   kable_classic_2(full_width = F, html_font = "Cambria") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))
```

## Score means by dimension

```{r}
# Find means across domains but within each dimension
dimension_mean <- attributes_merged_clean %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list("Mean" = mean, "Standard Deviaton" = sd)) %>% 
  rename(Dimension = dimension) %>% 
  mutate(across(where(is.numeric), ~ round(., 3)))

# Create a table
dimension_mean %>% 
  kbl(caption = "Attribute Score Means by Dimension") %>% 
   kable_classic_2(full_width = F, html_font = "Cambria") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))
```

## Score means by domain

```{r}
# Find means across domains but within each dimension
domain_mean <- attributes_merged_clean %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Dimension = dimension,
         Domain = domain,
         Score = score) %>% 
  mutate(across(where(is.numeric), ~ round(., 3)))

# Create a table
domain_mean %>% 
  kbl(caption = "Attribute Score Means by Domain") %>% 
   kable_classic_2(full_width = F, html_font = "Cambria") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))
```



##############################################################################################

# Grouping Attribute Score Stats

## Attribute Score Means by Dimension
```{r}
# Create a table where the average scores of each domain across case studies are represented

# Aguion mean
aguion_dim <- attributes_merged_clean %>% 
  filter(case_study == "Aguion") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Aguion = score,
         Dimension=dimension) 

# Go back to Burden case in the 01-prep-data RMD and remove filter NA for data quality score
# Burden mean
burden_dim <- attributes_merged_clean %>% 
  filter(case_study == "Burden") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Burden = score)

# Dickey-Collas mean
dickey_collas_dim <- attributes_merged_clean %>% 
  filter(case_study == "Dickey_Collas") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Dickey_Collas = score) 

# Eurich mean
eurich_dim <- attributes_merged_clean %>% 
  filter(case_study == "Eurich") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Eurich = score)

# Free mean
free_dim <- attributes_merged_clean %>% 
  filter(case_study == "Free") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Free = score)

# Friedman mean
friedman_dim <- attributes_merged_clean %>% 
  filter(case_study == "Friedman") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Friedman = score)

# Golden mean
golden_dim <- attributes_merged_clean %>% 
  filter(case_study == "Golden") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Golden = score)

# Hollowed mean
hollowed_dim <- attributes_merged_clean %>% 
  filter(case_study == "Hollowed") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Hollowed = score)

# Kisara mean
kisara_dim <- attributes_merged_clean %>% 
  filter(case_study == "Kisara") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Kisara = score)

# Kleisner mean
kleisner_dim <- attributes_merged_clean %>% 
  filter(case_study == "Kleisner") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Kleisner = score)

# Lau mean
lau_dim <- attributes_merged_clean %>% 
  filter(case_study == "Lau") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Lau = score)

# Mason mean
mason_dim <- attributes_merged_clean %>% 
  filter(case_study == "Mason") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Mason = score)

# Mills mean
mills_dim <- attributes_merged_clean %>% 
  filter(case_study == "Mills") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Mills = score)

# Pecl mean
pecl_dim <- attributes_merged_clean %>% 
  filter(case_study == "Pecl") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Pecl = score)

# Schmidt mean
#schmidt_dim <- attributes_merged_clean %>% 
#  filter(case_study == "Schmidt") %>% 
#  group_by(dimension) %>% 
#  summarise_at(vars(official_score), list(score = mean)) %>% 
#  rename(Schmidt = score)

# Tokunaga mean
tokunaga_dim <- attributes_merged_clean %>% 
  filter(case_study == "Tokunaga") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Tokunaga = score)

# Westfall mean
westfall_dim <- attributes_merged_clean %>% 
  filter(case_study == "Westfall") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Westfall = score)

# Zhao mean
zhao_dim <- attributes_merged_clean %>% 
  filter(case_study == "Zhao") %>% 
  group_by(dimension) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Zhao = score)
```

```{r}
### Kisara Case Study missing b/c no information for some of the dimensions 

# Merge all mean datasets together
dimension_means_merged <- cbind(aguion_dim, burden_dim, dickey_collas_dim, eurich_dim, free_dim, friedman_dim, golden_dim, hollowed_dim,  kleisner_dim, lau_dim, mason_dim, mills_dim, pecl_dim, tokunaga_dim, westfall_dim, zhao_dim) %>% 
  select(Dimension, Aguion, Burden, Dickey_Collas, Eurich, Free, Friedman, Golden, Hollowed, Kleisner, Lau, Mason, Mills, Pecl, Tokunaga, Westfall, Zhao) %>% 
  mutate(across(where(is.numeric), ~ round(., 3)))


# Create a table
dimension_means_merged %>% 
  kbl(caption = "Attribute Score Means Across by Dimension") %>% 
   kable_classic_2(full_width = F, html_font = "Cambria") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))
```


## Attribute Score Means by Domain

```{r}
# Create a table where the average scores of each domain across case studies are represented

# Aguion mean
aguion_mean <- attributes_merged_clean %>% 
  filter(case_study == "Aguion") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Aguion = score) 

# Burden mean
burden_mean <- attributes_merged_clean %>% 
  filter(case_study == "Burden") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Burden = score)

# Dickey-Collas mean
dickey_collas_mean <- attributes_merged_clean %>% 
  filter(case_study == "Dickey_Collas") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Dickey_Collas = score) 

# Eurich mean
eurich_mean <- attributes_merged_clean %>% 
  filter(case_study == "Eurich") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Eurich = score)

# Free mean
free_mean <- attributes_merged_clean %>% 
  filter(case_study == "Free") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Free = score)

# Friedman mean
friedman_mean <- attributes_merged_clean %>% 
  filter(case_study == "Friedman") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Friedman = score)

# Golden mean
golden_mean <- attributes_merged_clean %>% 
  filter(case_study == "Golden") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Golden = score)

# Hollowed mean
hollowed_mean <- attributes_merged_clean %>% 
  filter(case_study == "Hollowed") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Hollowed = score)

# Kisara mean
kisara_mean <- attributes_merged_clean %>% 
  filter(case_study == "Kisara") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Kisara = score)

# Kleisner mean
kleisner_mean <- attributes_merged_clean %>% 
  filter(case_study == "Kleisner") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Kleisner = score)

# Lau mean
lau_mean <- attributes_merged_clean %>% 
  filter(case_study == "Lau") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Lau = score)

# Mason mean
mason_mean <- attributes_merged_clean %>% 
  filter(case_study == "Mason") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Mason = score)

# Mills mean
mills_mean <- attributes_merged_clean %>% 
  filter(case_study == "Mills") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Mills = score)

# Pecl mean
pecl_mean <- attributes_merged_clean %>% 
  filter(case_study == "Pecl") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Pecl = score)

# Tokunaga mean
tokunaga_mean <- attributes_merged_clean %>% 
  filter(case_study == "Tokunaga") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Tokunaga = score)

# Westfall mean
westfall_mean <- attributes_merged_clean %>% 
  filter(case_study == "Westfall") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Westfall = score)

# Zhao mean
zhao_mean <- attributes_merged_clean %>% 
  filter(case_study == "Zhao") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = mean)) %>% 
  rename(Zhao = score)
```

```{r}
### Kisara Case Study missing b/c no information for some of the dimensions 

# Merge all mean datasets together
means_merged_by_case <- cbind(aguion_mean, burden_mean, dickey_collas_mean, eurich_mean, free_mean, friedman_mean, golden_mean, hollowed_mean,  kleisner_mean, lau_mean, mason_mean, mills_mean, pecl_mean, tokunaga_mean, westfall_mean, zhao_mean) %>% 
  select(dimension...1, domain...2, Aguion, Burden, Dickey_Collas, Eurich, Free, Golden, Friedman, Hollowed, Kleisner, Lau, Mason, Mills, Pecl, Tokunaga, Westfall, Zhao) %>% 
  rename(Dimension = dimension...1, Domain = domain...2) %>% 
  mutate(across(where(is.numeric), ~ round(., 2)))

# Create a table
means_merged_by_case %>% 
  kbl(caption = "Resilience Attribute Score Means Across Case Studies") %>% 
   kable_classic_2(full_width = F, html_font = "Cambria") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))
```


####################################################################################################

# Summing the attributes by dimension and domain

```{r}
# Create a table where the sums if each attribute score are grouped by dimension and domain

# Aguion sum
aguion_sum <- attributes_merged_clean %>% 
  filter(case_study == "Aguion") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Aguion = score) 

# Burden sum
burden_sum <- attributes_merged_clean %>% 
  filter(case_study == "Burden") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Burden = score)

# Dickey-Collas sum
dickey_collas_sum <- attributes_merged_clean %>% 
  filter(case_study == "Dickey_Collas") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Dickey_Collas = score) 

# Eurich sum
eurich_sum <- attributes_merged_clean %>% 
  filter(case_study == "Eurich") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Eurich = score)

# Free sum
free_sum <- attributes_merged_clean %>% 
  filter(case_study == "Free") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Free = score)

# Friedman sum
friedman_sum <- attributes_merged_clean %>% 
  filter(case_study == "Friedman") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Friedman = score)

# Golden sum
golden_sum <- attributes_merged_clean %>% 
  filter(case_study == "Golden") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Golden = score)

# Hollowed sum
hollowed_sum <- attributes_merged_clean %>% 
  filter(case_study == "Hollowed") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Hollowed = score)

# Kisara mean
#kisara_mean <- attributes_merged_clean %>% 
#  filter(case_study == "Kisara") %>% 
#  group_by(dimension, domain) %>% 
#  summarise_at(vars(official_score), list(score = mean)) %>% 
#  rename(Kisara = score)

# Kleisner sum
kleisner_sum <- attributes_merged_clean %>% 
  filter(case_study == "Kleisner") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Kleisner = score)

# Lau sum
lau_sum <- attributes_merged_clean %>% 
  filter(case_study == "Lau") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Lau = score)

# Mason sum
mason_sum <- attributes_merged_clean %>% 
  filter(case_study == "Mason") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Mason = score)

# Mills sum
mills_sum <- attributes_merged_clean %>% 
  filter(case_study == "Mills") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Mills = score)

# Pecl sum
pecl_sum <- attributes_merged_clean %>% 
  filter(case_study == "Pecl") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Pecl = score)

# Tokunaga sum
tokunaga_sum <- attributes_merged_clean %>% 
  filter(case_study == "Tokunaga") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Tokunaga = score)

# Westfall sum
westfall_sum <- attributes_merged_clean %>% 
  filter(case_study == "Westfall") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Westfall = score)

# Zhao sum
zhao_sum <- attributes_merged_clean %>% 
  filter(case_study == "Zhao") %>% 
  group_by(dimension, domain) %>% 
  summarise_at(vars(official_score), list(score = sum)) %>% 
  rename(Zhao = score)
```

```{r}
### Kisara Case Study missing b/c no information for some of the dimensions 

# Merge all sum datasets together
sums_merged_by_case <- cbind(aguion_sum, burden_sum, dickey_collas_sum, eurich_sum, free_sum, friedman_sum, golden_sum, hollowed_sum,  kleisner_sum, lau_sum, mason_sum, mills_sum, pecl_sum, tokunaga_sum, westfall_sum, zhao_sum) %>% 
  select(dimension...1, domain...2, Aguion, Burden, Dickey_Collas, Eurich, Free, Golden, Friedman, Hollowed, Kleisner, Lau, Mason, Mills, Pecl, Tokunaga, Westfall, Zhao) %>% 
  rename(Dimension = dimension...1, Domain = domain...2) %>% 
  mutate(across(where(is.numeric), ~ round(., 2)))

# Create a table
sums_merged_by_case %>% 
  kbl(caption = "Resilience Attribute Score Sums Grouped by Dimension and Domain Across Case Studies") %>% 
   kable_classic_2(full_width = F, html_font = "Cambria") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))
```
####################################################################################################

# Visualize Attribute Score Means

```{r}
# Create theme
my_theme <-  theme(axis.text.x=element_text(size=7, angle=38, hjust=0.8),
                   axis.text.y=element_text(size=8),
                   legend.text=element_text(size=8),
                   legend.title=element_text(size=9),
                   strip.text=element_text(size=9),
                   # Gridlines
                   panel.grid.major = element_blank(), 
                   panel.grid.minor = element_blank(),
                   panel.background = element_rect(), 
                   axis.line = element_line())
```

```{r}
# Bar graph showing average comparison by dimension across case studies

dimension_comparison <- dimension_means_merged %>% 
  pivot_longer(!Dimension, names_to = "case_study", values_to = "means") %>% 
  mutate(across(where(is.numeric), ~ round(., 2)))


score_dimension_graph <- ggplot(dimension_comparison, aes(x = case_study, y = means, fill = Dimension)) +
  geom_bar(stat="identity", width=0.7, position=position_dodge(width=1)) +
  theme_bw() + my_theme +
  facet_wrap(~Dimension, scales = "free") +
  theme(legend.position="top") +
  geom_hline(data=dimension_mean, aes(yintercept = (Mean)), linetype="dashed", colour="black") +
  labs(x = "Case Studies", y = "Attribute Score Averages") +
  scale_fill_brewer(palette = "Accent")
plot(score_dimension_graph)
```

```{r}
# Make a stacked graph for the data

score_stacked <- ggplot(dimension_comparison, aes(x = case_study, y = means, fill = Dimension, label = means)) +
  geom_bar(stat="identity", position="stack") +
  theme_bw() + my_theme +
  theme(legend.position="top") +
  labs(x = "Case Studies", y = "Attribute Score Averages") +
  scale_fill_brewer(palette = "Accent") +
  geom_text(size = 3, position = position_stack(vjust = 0.5))
plot(score_stacked)
```

# Visualize sum

```{r}
# Create dataset

sum_comparison <- sums_merged_by_case %>% 
  pivot_longer(cols = c("Aguion", "Burden", "Dickey_Collas", "Eurich", "Free", "Golden", "Friedman", "Hollowed", "Kleisner", "Lau", "Mason", "Mills", "Pecl", "Tokunaga", "Westfall", "Zhao"),
               names_to = "case_study", 
               values_to = "sums") %>% 
  mutate(across(where(is.numeric), ~ round(., 2))) %>% 
  mutate()
# need to break it into dimensions - plot eahc one individually


# Make a stacked graph for the data

score_sum_stacked <- ggplot(sum_comparison, aes(x = case_study, y = sums, fill = Domain, label = sums)) +
  geom_bar(stat="identity", position="stack") +
  theme_bw() + my_theme +
  theme(legend.position="top") +
  labs(x = "Case Studies", y = "Attribute Score Averages") +
  scale_fill_brewer(palette = "Accent") +
 # geom_text(size = 3, position = position_stack(vjust = 0.5)) +
  facet_wrap(~Dimension, scales="free")
plot(score_sum_stacked)
```

##################################################################################################

# Attribute Score Proportions

```{r}
# Create a table that shows the counts and proportions of attribute scores to help identify which are scoring higher than others and vice versa.

attribute_score_stats <- attributes_merged_clean %>% 
  select(dimension, score) %>% 
  tabyl(dimension,score) %>%
  adorn_percentages() %>% 
  adorn_pct_formatting(digits = 2) %>% 
  column_to_rownames(var = "dimension")


# Create a table
attribute_score_stats %>% 
  kbl(caption = "Proportion of Scores by each Dimension") %>% 
   kable_classic_2(full_width = F, html_font = "Cambria") %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))
```
